requirements.txt
streamlit>=1.45.1,<2.0.0
openai-whisper>=20240930